--- webapps/ui/cockpit/client/scripts/components/callActivityDefinitionOverlay/template.html
@@ -1,3 +1,0 @@
-<button type="button" role="button" class="btn btn-sm btn-default call-btn activity-top-right-position">
-  <span class="glyphicon glyphicon glyphicon-link"></span>
-</button>
